+++ source/node.h
@@ -23,7 +23,7 @@ namespace k_tree
 		friend class k_tree;
 
 		private:
-			static constexpr float float_resolution = 0.000001;														// floats his close are considered equal
+			static constexpr float float_resolution = (float)0.000001;														// floats his close are considered equal
 
 		public:
 			size_t max_children;					//	the order of the tree at this node (constant per tree as it propegates when a new node is created)
